back/requirements.txt
starlette
numpy
pandas
pydantic
opencv-python
requests
supervision
transformers
pytorch_lightning
fastapi